KUDU-3413 add tenant info in metadata for multi-tenancy
If we want to introduce multi-tenancy feature, it is important to
add relevant metadata information first. In this patch, I added
definition of tenant information in metadata, and refactored some
metadata loading to distinguish between data at rest encryption
feature and multi-tenancy feature.

In the subsequent use, if we encounter a situation where both server
key and tenant key are present in the metadata, we need to pay
attention to the fact that tenant key will have a higher priority.
That is, we will consider it as enabling the multi-tenancy feature
and ignore the processing of server key.

Of course, it is also necessary to provide a way to upgrade server key
to tenant key, which I plan to implement in the next patch.

In this commit, I added a new flag '--enable_multi_tenancy' to indicate
whether we enable multi-tenancy feature. However, considering that the
development of the entire feature has not been completed, it is not
recommended to set this flag temporarily. We need to wait until the
entire feature is developed before considering whether to set it.

And I also added unit tests for the new open process to ensure the
open logic is effective.

Change-Id: I9e450d73940eb1dbaac6f905a46d6ccd084f15cf
Reviewed-on: http://gerrit.cloudera.org:8080/19622
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>