[master] Exclude tservers in MAINTENANCE_MODE when leader rebalancing
Currently, auto leader rebalancing may transfer new leaders to tservers
which have entered MAINTENANCE_MODE. Because a tserver in MAINTENANCE_MODE
is being maintained by administrators, maybe it is decommissioning.
Since it might be decommissioned/decommissioning, it doesn't make sense
for the leader rebalancer to transfer a leader to a tserver in MAINTENANCE_MODE
as leaders serve many users' requests. For this reason, we should exclude
such tservers.

This patch fixes the problem and adds an unit test for this.

Change-Id: I2f85a675e69fd02a62e2625881dad2ca5e27acd9
Reviewed-on: http://gerrit.cloudera.org:8080/19608
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Yuqi Du <shenxingwuying@gmail.com>
Reviewed-on: http://gerrit.cloudera.org:8080/19923
Tested-by: Yingchun Lai <laiyingchun@apache.org>