KUDU-3448 Plumbing for encrypting key material
Key material for the internal PKI and token signing keys are stored in
the syscatalog table in clear text, which is okay when volume-level
encryption or Kudu's built-in data at rest encryption is used, but in
some cases, this is either not used, or it's not enough (FISMA).

To allow storing these key materials in encrypted form in the syscatalog
table, this patch adds the necessary plumbing in Kudu's OpenSSL wrapper.
It is now possible to pass a password callback function to the utility
functions responsible for reading from and writing to OpenSSL BIO and
strings.

Change-Id: I24c5ac8ea0f9a4cab0f35ecccb1b7b00f3acefa8
Reviewed-on: http://gerrit.cloudera.org:8080/19615
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>