KUDU-3448 Add support for encrypting IPKI keys
This patch introduces a new flag, --ipki_private_key_password_cmd. If
set, Kudu's internal PKI's root CA private key will be encrypted with
the password that is output by the command set with this flag.

The key is encrypted with AES-256-CBC and encoded in PKCS#8 format. The
behavior is similar to --webserver_private_key_password_cmd, which is
used to provide a command to decrypt the webserver certificate's private
key.

Currently, Kudu doesn't support rotating IPKI keys, so this flag can't
be used on existing clusters, and if it was used on the first startup of
a master, it must be used as long as that master exists, it won't be
able to start without it.

Change-Id: I71f2ec856f018d56efbf6901039eed2676fcbe23
Reviewed-on: http://gerrit.cloudera.org:8080/19616
Reviewed-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Tested-by: Kudu Jenkins