[Python] Refactor tests to use assertRaisesRegex
The Python unittest.TestCase class offers the assertRaises() and
assertRaisesRegex() functions to test that an exception is raised when
callable is called [1].

Currently assertRaises(exception) is used in the Python
client tests, to assert the type of a given error. This patch refactors
usages of assertRaises() to assertRaisesRegex(exception, regex) in order
to assert the error message as well.

In Python3.2 assertRaisesRegexp has been renamed to assertRaisesRegex
(without the trailing 'p') [2].  To be able to use assertRaisesRegex
throughout the Python codebase, a compatibility class called
CompatUnitTest is added.

[1] https://docs.python.org/3/library/unittest.html#unittest.TestCase
[2] https://docs.python.org/3.9/library/unittest.html
    #unittest.TestCase.assertRaisesRegex

Change-Id: I41e2d69996ee0ed0f0418ae184d95239f2739efb
Reviewed-on: http://gerrit.cloudera.org:8080/19633
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>
(cherry picked from commit 08bd36a8745d20c5aa688f5c33a05615857d40e0)
Reviewed-on: http://gerrit.cloudera.org:8080/19981
Reviewed-by: Marton Greber <greber.mx@gmail.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>