Change order of operations during InitCatalogManagerTask
In MasterStartupTest.StartupWebPage test, WaitForCatalogManagerInit is
executed, which waits until catalog_manager_init_status_ is set. It is
set by InitCatalogManagerTask and with the previous order of operations
it was possible (but very rare) that catalog_manager_init_status_ was
set but the timer was not stopped at the time of the startup http
request, which resulted an output where
initialize_master_catalog_status was 0. With this order
initialize_master_catalog_progress_ is always set if
WaitForCatalogManagerInit is finished.

Change-Id: I64c980a8b357c99d4431a02d58b3c03a60bae1ad
Reviewed-on: http://gerrit.cloudera.org:8080/19639
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>