[rowset_metadata] update min/max encoded keys during bootstrapping
We can cache the encoded min/max keys in rowset metadata to help
tablets to bootstrap without having to fully initialize the CFileReaders
for the key columns of each rowset. However, we can only update the
min/max encoded keys during insert or update operations right now.
This means that if we didn't save the min/max encoded key before,
we won't have the opportunity to use this feature to accelerate bootstrap
in the future.

Considering that we will read the min/max key values during the bootstrap
process, we can set the min/max encoded keys in the metadata at the same
time.

I added a small test case to ensure we can get the same keys from the bootstrap
process at startup.

Change-Id: I1b7ed75bdf7a2ff0e16c2670f1a6f9819ee8e8d3
Reviewed-on: http://gerrit.cloudera.org:8080/19650
Tested-by: Kudu Jenkins
Reviewed-by: Yuqi Du <shenxingwuying@gmail.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>