[jwt] Verify JWKS URL server TLS certificate by default
This commit is to pull IMPALA-11922 code into the Kudu jwt handling,
with some modifications.

This change introduces:
  1. verification of JWKS server TLS certificate by default
  2. jwks_verify_server_certificate Kudu startup flag

Instead of introducing a new flag such as 'jwks_ca_certificate' the
already existing 'trusted_certificate_file' flag is reused.

The TLS certificate verification is not used in unit-tests, however
security-itest is set up with the verification enabled.

Change-Id: I0fd7b53d651786bbe57642dd14cd477055b80c78
Reviewed-on: http://gerrit.cloudera.org:8080/19709
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Kudu Jenkins