jwt: Java client usage of JWTs
This change introduces modifications to the Negotiator and the
SecurityContext  classes to be able to handle JWT authentication from
the Java client.

Also adding a test where the client is set up with a JWT that was
generated by the MiniOIDC that starts along with the
ExternalMiniCluster.

Change-Id: Id02dcd7ee57a838f4763500e78053e21aac21c09
Reviewed-on: http://gerrit.cloudera.org:8080/18477
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Attila Bukor <abukor@apache.org>
(cherry picked from commit a3a7c97be031f8fc32402e430eff1a89c19dbdfb)
Reviewed-on: http://gerrit.cloudera.org:8080/19972
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>