[build] Update libraries to be excluded in SLES15 SP4
Recently during the build of minicluster from
build-support/mini-cluster/build_mini_cluster_binaries.sh
we found that libcurl built from third party is dependent on the
library libjitterentropy. While invesitgating if this dependency is
present just because libjitterentropy is present in the machine and
curl picking it up while  building, we found that this library
libjitterentropy is present in SLES15 SP4 out of the box.
We also found having libjitterentropy in a RHEL 9 machine didn't cause
the dependency behavior from libcurl i.e. it is not dependent on
libjitterentropy.
To avoid these build issues in SELS 15 SP4, libjitterentropy needs to
be excluded when copying the libraries to be distributed. On a side
note libjitterentropy is also BSD and GPLv2 licensed.

Change-Id: I8152d7baaa944e1573c076b36980c73902c426ca
Reviewed-on: http://gerrit.cloudera.org:8080/19749
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>