[build-support] allow building with gcc/g++-7 on SLES
SLES15 has GCC7 as the system compiler, so Kudu can be built with it
instead of installing GCC8 packages from the Development Tools Module.
If both GCC7 and GCC8 are present, prefer GCC8 over GCC7 to build Kudu.

This patch also updates the enable_devtoolset.sh script to properly
handle the absence of the `lsb_release` binary.

Change-Id: I4fd08cfb1cdc336b44b9568cb4dc019575f01efa
Reviewed-on: http://gerrit.cloudera.org:8080/19778
Tested-by: Kudu Jenkins
Reviewed-by: Ashwani Raina <araina@cloudera.com>
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>