[util] use include guard macro in int128.h header
Even if '#pragma once' is supported by the vast majority of modern
compilers, it is a non-standard one [1].  The src/kudu/util/int128.h
file is included into other header files of the C++ Kudu client API,
which is supposed to be C++98 compliant.  Since other header files
in the C++ client API are using include guard macros, let's use
the same technique for int128.h as well.

Doing so addresses a warning produced by the CLANG when running
client_examples_test.sh, while a standalone compilation is done on the
kudu/util/int128.h file:

  #pragma once in main file [-Wpragma-once-outside-header]

[1] https://en.cppreference.com/w/cpp/preprocessor/impl

Change-Id: Ica18bbdc8eff034965e1a69404df2180cacf7bb4
Reviewed-on: http://gerrit.cloudera.org:8080/19783
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>