[tests] fix flakiness in TestNoMoreRetryWithWongServerUuid
The TestNoMoreRetryWithWongServerUuid scenario of the DeleteTabletITest
was a bit flaky.  The system catalog was sometimes sending in
DeleteTablet RPCs when the original tablet server was still starting
up, so even if DeleteTablet requests went through and were accounted
for, they were responded with an error status because the tablet
manager of the original tablet server wasn't ready to receive such
requests yet, and the catalog manager retried its requests after
receiving such responses.

This patch addresses the flakiness and fixes a typo in the name of the
scenario.  The scenario's dist-test stats for a DEBUG build are below.

Before:
  6 out of 1024 failed:
    http://dist-test.cloudera.org/job?job_id=aserbin.1682129573.9602

After:
  0 out of 1024 failed:
    http://dist-test.cloudera.org/job?job_id=aserbin.1682133036.42291

Change-Id: I3d3dee89b32d1e33d1f0f41e8b83835b02eae336
Reviewed-on: http://gerrit.cloudera.org:8080/19785
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-on: http://gerrit.cloudera.org:8080/19947
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Yuqi Du <shenxingwuying@gmail.com>
Reviewed-by: Yifan Zhang <chinazhangyifan@163.com>