[client] KUDU-3472 C++ client API to import JWT
This patch addresses add C++ client API to specify a JWT for a newly
built C++ client instance.  It's an improvement in the usability of the
Kudu client API since with this new extra method the users don't need
to mess with the serialization of the authn credentials.

The new functionality is covered by already existing tests that have
been updated to start using the new API.  More tests might be added
to cover other aspects of the JWT-related functionality, but these
are not directly related to this API change, so we can take care of them
in a separate patch.

Change-Id: I8333f1d0c3b4e92ef8430c7ad91bb7da963aceb9
Reviewed-on: http://gerrit.cloudera.org:8080/19808
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>
Tested-by: Alexey Serbin <alexey@apache.org>