[tests] Fix flakiness in log-rolling-itest.cc
The test TestLogCleanupOnStartup within log-rolling-itest.cc
has been flaky with a 30% failure rate for the last 2+ weeks.
The timeout within the test was increased from 30 to 60 seconds
to allow the ExcessLogFileDeleterThread to have enough time to
properly clean up the log files.

Below are the dist-test stats for a DEBUG build.

Before: 13/100 failed.
http://dist-test.cloudera.org/job?job_id=mreddy.1683141191.148826

After: 0/100 failed.
http://dist-test.cloudera.org/job?job_id=mreddy.1683146152.19719

Change-Id: I4d33559fa01c26d8c8b3ad12b8822d2cdd760fae
Reviewed-on: http://gerrit.cloudera.org:8080/19842
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>