KUDU-3474: Add zlib as dependency of libunwind for ARM
On ARM, libunwind includes support for compressed
.debug_info sections if zlib is available during
compilation. This means that libunwind can require zlib
at link time, but only for ARM architecture.

This modifies the build to add a zlib dependency for
libunwind if the architecture is ARM. It also reorders
the thirdparty build so that zlib is built before
libunwind. This is not strictly necessary, but it means
that libunwind would consistently have support for
compressed .debug_info sections on ARM. This does not
impact x86_64.

Testing:
 - Ran an ARM build on Ubuntu 20.04

Change-Id: Iee87b908a5771c2bd7c1a653504a1449d0792280
Reviewed-on: http://gerrit.cloudera.org:8080/19846
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit 96e648ac4a3d80a5d2194c834d82035fc90039e4)
Reviewed-on: http://gerrit.cloudera.org:8080/20051
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>