KUDU-3448 Move password retrieval to master init
Apparently, forks can be expensive, at least on some systems, which
makes fetching the password to encrypt/decrypt key materials whenever a
TSK or IPKI is generated or loaded into memory not ideal as it may cause
the server process to hang.

This patch moves the password retrieval to the master initialization
phase and the passwords are now stored in memory for later use.

Change-Id: I746e657ae8d295f5f34225d63686beea1dff6b7c
Reviewed-on: http://gerrit.cloudera.org:8080/19873
Tested-by: Attila Bukor <abukor@apache.org>
Reviewed-by: Alexey Serbin <alexey@apache.org>