[master] add /ipki-ca-cert webserver end-point
This patch introduces a new /ipki-ca-cert end-point for the master's
embedded webserver to provide information on the IPKI CA certificate
used by a Kudu cluster.  This information is useful in JWT-based
authentication scenarios.  The missing part of importing certificates
into Kudu client's chain of trusted TLS certificates will be addressed
in follow-up patches.

Change-Id: I833d934cc1d916dba243a05aa96926d3b540b70d
Reviewed-on: http://gerrit.cloudera.org:8080/19906
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>