[jwt] switching JWT verification to KeyBasedJwtVerifier
Switching JWT verification to use KeyBasedJwtVerifier instead of
PerAccountKeyBasedJwtVerifier, as the latter relies on having an OIDC
discovery endpoint available even if a jwks_url is set. Fixing the jwks
discovery via an OIDC server will be done in a later patch.

This change also moves the initialisation of the JWTVerifier from
startup to the VerifyToken() method. In order to test this a new option
'start_jwks' for the ExternalMiniCluster was introduced.

Change-Id: Ic1f166807bfcf7051bda7843e186eacfbe379eba
Reviewed-on: http://gerrit.cloudera.org:8080/19910
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <alexey@apache.org>