[java] allow adding CA certs to the client's trust chain
With this patch, the Java client doesn't implicitly trust the TLS
certificate of a Kudu server it's negotiating an RPC connection with
when using its JWT. There is also an update on the Negotiator class
to fix various JWT-related issues during RPC connection negotiation.

To implement that, this patch introduces the following new methods:

  * [Async]KuduClient.trustedCertificates(List<ByteString> certs)
    to add the specified CA certificates (in DER format) as the
    trusted ones for the client (the call replaces all the previously
    set certificates if invoked again)

  * KuduTestHarness.getClusterCACertDer() to acquire the IPKI CA
    certificate of the test mini-cluster via call to
    MiniKuduCluster.getCACertDer()

  * MiniKuduCluster.getMasterWebServerAddressesAsString() which
    returns the addresses of Master's webserver as a comma separated
    string, similar to MiniKuduCluster.getMasterAddressesAsString()

In addition, this changelist also contains tests to cover the newly
introduced functionality.

Change-Id: I281fd59da0d24a6119b99e1da8096f381b414a2b
Reviewed-on: http://gerrit.cloudera.org:8080/20017
Tested-by: Kudu Jenkins
Reviewed-by: Attila Bukor <abukor@apache.org>