[webserver] handlers for application/octet-stream content
This patch introduces application/octet-stream (a.k.a. binary data)
handlers for the embedded web server.

As a particular application of the newly introduced feature, this patch
also adds /ipki-ca-cert-der endpoint to the master's embedded webserver.
It outputs the same IPKI CA certificate as the /ipki-ca-cert and
/ipki-ca-cert-pem endpoints, but in DER format.  This will be used
in a changelist where Java client needs to import Kudu cluster's CA
certificate into its trust chain when using JWT for authentication.
As it turns out, the 'standard' security providers for Java runtime
don't provide enough functionality to conveniently work with X509 in PEM
format (yes, there is BouncyCastle, but I'm not sure we want to
introduce such a dependency just for the convenience of working with
certificates in PEM format at a few call sites in the Java client).

To cover the newly introduced functionality, an extra test has been added
into security-itest and PeriodicWebUIChecker has been updated.

Change-Id: I894d8e00943617cb80ec5aa14a15db3448ad9252
Reviewed-on: http://gerrit.cloudera.org:8080/20023
Tested-by: Kudu Jenkins
Reviewed-by: Zoltan Chovan <zchovan@cloudera.com>
Reviewed-by: Attila Bukor <abukor@apache.org>