[cfile] replace BlockPointer::Equals() with operator==()
The motivation for this change is a realisation that methods like
Equals() look a bit funny in C++ code since the language supports
much more expressive and readable notation for comparing objects.

I also did a minor cleanup to bring the code in block_pointer.h
to the style guide and removed unused instances of the BlockPointer
class elsewhere in the code.

This patch doesn't contain any functional changes.

Change-Id: I3b2c05931ce4b678ef184d0c812cf0fb51b650c7
Reviewed-on: http://gerrit.cloudera.org:8080/20031
Tested-by: Alexey Serbin <alexey@apache.org>
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>