KUDU-1945 Backup/restore for tables with auto incrementing columns
This patch adds the ability to backup and restore tables containing
auto-incrementing columns. A simple test case is also added to test
the functionaility.

Without this patch any attempt to
Backup a table with auto-incrementing column will fail with an error
message "auto_incrementing_id is a reserved column name" from
schema.cc.
Restoring a table with auto-incrementing column will result in the
same behavior as with the backup.

Change-Id: I929d54d22c1c938ee67fdda9f4c2bb68c028b4ec
Reviewed-on: http://gerrit.cloudera.org:8080/20084
Reviewed-by: Alexey Serbin <alexey@apache.org>
Tested-by: Kudu Jenkins