[util] fix at-rest {en,de}cryption for OpenSSL 3
It seems that with OpenSSL 3.0 padding is enabled by default [1].
However, the code in env-posix.cc was relying on an incorrect assumption
that the padding would be always disabled for AES-CTR ciphers, and there
were no calls to EVP_{En,De}cryptFinal_ex() in Do{En,De}cryptV()
correspondingly.

This patch addresses the issue, disabling padding for the cipher
context explicitly. I also updated the order of the expected/actual
arguments for ASSERT_EQ() because the error message was a bit hard
to read and comprehend if the corresponding test failed.

The motivation for this patch were reports on failed env-test if built
and run on RHEL9.

Probably, adding calls to EVP_{En,De}cryptFinal_ex() into
Do{En,De}cryptV() and making sure they return nothing would be a good
invariant to check in DEBUG builds.  I added corresponding TODOs in the
code since it's better to address that in a separate patch.

This is a follow-up to de02a34390f9bdb722be39f96ef9ad9573c0eeeb.

[1] https://www.openssl.org/docs/man3.0/man3/EVP_EncryptUpdate.html

Change-Id: Ia49a51488a0a39953249f24598daa8219583d575
Reviewed-on: http://gerrit.cloudera.org:8080/20100
Reviewed-by: Attila Bukor <abukor@apache.org>
Tested-by: Alexey Serbin <alexey@apache.org>
(cherry picked from commit ed1fe26c6d7dce49496db1750771963ae4a5f5fa)
Reviewed-on: http://gerrit.cloudera.org:8080/20413
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>