[java] introduce skipCodeStyleChecks gradle property
It doesn't make much sense to run code style verification in various
non-development build environments since the code isn't supposed
to change and be pushed back into the repository in such a case.  All
those verification and checks are enabled by default to be run in
standard development environment and various pre-commit verification
builds.

With this patch, it's now possible to define the 'skipCodeStyleChecks'
property to avoid running those code style verifications for Java and
Scala code:

  ./gradlew -PskipCodeStyleChecks=true ...

Change-Id: I6e67e71ab21908166aba0f83a10b4aba0fdc5f3d
Reviewed-on: http://gerrit.cloudera.org:8080/20116
Tested-by: Kudu Jenkins
Reviewed-by: Abhishek Chennaka <achennaka@cloudera.com>