[multi-tenancy] add add/remove tenant api in fs layer
In this patch, my main addition is the API for adding and
deleting tenants in the fs layer.

Additionally, I adjusted the env, dd_manager, block_manager,
and other relevant parameters to be tenant-related. For the
default tenant, which is used for both enabled and not enabled
multi-tenant features, the default parameters will be used.
New tenants enabled with multi-tenant feature will possess their
own unique environment parameters.

Also, I added unit tests to ensure the validity of the newly added
logic.

However, this patch leaves the standardization work for tenant
storage paths in multi-tenant scenarios to the next patch, where
I will focus on implementing this part.

Change-Id: I3f31d3ddd636952f8bd432330afbde018169a2a1
Reviewed-on: http://gerrit.cloudera.org:8080/20144
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>