[multi-tenancy] standardize the use of tenant's api
In order to standardize the use of tenants, in this patch,
I changed the defining parameter of multiple tenant-related
APIs from the tenant name to the tenant ID. In the future,
the update of tenant names will be supported and modification
of tenant IDs will be prohibited.

Since there are no logical changes involved, no new unit tests
have been added.

Change-Id: I5d8c09668456bc1147d8e269ecaa8e616798d2c7
Reviewed-on: http://gerrit.cloudera.org:8080/20157
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>
Tested-by: Yingchun Lai <laiyingchun@apache.org>