KUDU-3326 disable compaction for soft-deleted table
The tables in the 'SOFT_DELETED' state will no longer
process insert and update requests, so the compaction
and other related operations will become redundant and
also waste resources.

In this patch, we added a post-operation to disable
compaction for the soft-deleted operation and a
post-operation to enable compaction for the recall
process.

To verify the correctness of the new logic, we also
added new unit tests.

Change-Id: I060810051613a19e6f5c6506effda2d698528839
Reviewed-on: http://gerrit.cloudera.org:8080/20178
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <laiyingchun@apache.org>